+++ src/components/general/ContactInformationContainer.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import '../styles/sponsors.css';
+import '../../styles/sponsors.css';
 
 interface ContactInformationContainerProps {
     href?: string
