+++ src/styles/app.css
@@ -37,6 +37,7 @@
     flex-direction: row;
     align-items: center;
     justify-content: left;
+    justify-self: center;
     
 }
