--- src/styles/projectcard.css
@@ -1,72 +1,0 @@
-
-.entireCard {
-    overflow: hidden;
-    box-shadow: 0 2px 20px gray;
-    border-radius: 10px;
-    display:flex;
-    flex-direction: column;
-    justify-content: space-between;
-    transition-property: all;
-    transition-duration: 0.5s;
-}
-
-.entireCard:hover {
-    scale:1.02;
-    transition-property: all;
-    transition-duration: 0.5s;
-}
-
-.cardImage {
-    height: 10rem;
-    width: 100%;
-    object-fit: cover;
-}
-
-.cardTitle {
-    text-align: center;
-}
-
-.buttonHolder {
-    text-align: center;
-    padding-bottom: 1rem;
-}
-
-.moreButton {
-    display: inline-block;
-    font-size: 20px;
-    color: rgb(0, 0, 0);
-    font-family: arial, sans-serif;
-    text-decoration: none;
-    
-    position: relative;
-    background-color: rgb(255, 255, 255);
-    
-    border-image: linear-gradient(135deg, rgb(198, 117, 255), rgb(0, 80, 199));
-    border-image-slice: 1;
-
-    /* -webkit-backface-visibility: hidden; */
-    z-index: 1;    
-    height:auto;
-}
-
-.moreButton:after {
-    position: absolute;
-    content: 'Learn More';
-    top: 0;
-    left: 0;
-    width: inherit;
-    height: inherit;
-    /* border-radius: inherit; */
-    background-image: linear-gradient(135deg, rgb(198, 117, 255), rgb(0, 80, 199));
-    transition: opacity 2s ease-out;
-    z-index: 2;
-    opacity: 0;
-    padding: inherit;
-    margin: inherit;
-    color:white;
-}
-
-.entireCard:hover .moreButton:after {
-    opacity: 1;
-}
-
